+++ src/AlpacaFillListener.cpp
@@ -6,7 +6,7 @@
 #include <optional>
 #include <thread>
 
-void pin_thread_to_core(std::thread &t, size_t core_id);
+void pin_thread_to_core(std::thread &t, uint32_t core_id);
 
 namespace {
